requirements.txt
# Django
Django==4.2.7
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1
django-redis==5.4.0

# Environment variables
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# Static files
whitenoise==6.6.0

# Image handling
Pillow==10.1.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production server
gunicorn==21.2.0

# Database
dj-database-url==2.1.0
